refactor(test_functions): migrate Forrester1D to YAML spec
Convert the Forrester1D test function from the legacy ABC-based system to a
YAML-driven spec as part of the Phase 2 fixed-dimension migration. The
function was previously named Forrester2008; the name is updated to follow
the established convention (e.g., Oakley1D).

- Rename `Forrester2008` to `Forrester1D`
- Add `forrester_1d.yaml` spec file
  with `Forrester2008` input spec
- Strip ABC boilerplate from `forrester_1d.py`,
  keeping only `evaluate()`
- Fix typos in the documentation
- Revise page title to descriptive-first form with attribution
- Fix matplotlib labels to use raw strings

Partially resolves: #503
Ref: #502
