+++ docs/fundamentals/optimization.md
@@ -18,11 +18,11 @@ kernelspec:
 The table below listed the available test functions typically used
 in the comparison of global optimization methods.
 
-|                           Name                            | Input Dimension |    Constructor    |
-|:---------------------------------------------------------:|:---------------:|:-----------------:|
-|           {ref}`Ackley <test-functions:ackley>`           |        M        |    `Ackley()`     |
-| {ref}`Forrester et al. (2008) <test-functions:forrester>` |        1        | `Forrester2008()` |
-|       {ref}`Rosenbrock <test-functions:rosenbrock>`       |        M        |  `Rosenbrock()`   |
+|                              Name                               | Input Dimension |    Constructor    |
+|:---------------------------------------------------------------:|:---------------:|:-----------------:|
+|              {ref}`Ackley <test-functions:ackley>`              |        M        |    `Ackley()`     |
+| {ref}`Forrester et al. (2008) 1D <test-functions:forrester-1d>` |        1        | `Forrester2008()` |
+|          {ref}`Rosenbrock <test-functions:rosenbrock>`          |        M        |  `Rosenbrock()`   |
 
 In a Python terminal, you can list all the available functions relevant
 for optimization applications using ``list_functions()`` and filter the results
