+++ src/uqtestfuns/test_functions/forrester_1d.yaml
@@ -1,0 +1,18 @@
+name: Forrester1D
+description: One-dimensional multimodal function from Forrester et al. (2008)
+tags:
+  - optimization
+  - metamodeling
+
+dimensions:
+  input: 1
+  output: 1
+
+inputs:
+  Forrester2008:
+    description: Search domain from Forrester et al. (2008)
+    marginals:
+      - name: x
+        distribution: uniform
+        parameters: [0.0, 1.0]
+    copulas: null
